requirements.txt
absl-py==0.1.13
astor==0.6.2
autopep8==1.3.5
backcall==0.1.0
bleach==1.5.0
boto==2.48.0
boto3==1.7.4
botocore==1.10.4
bz2file==0.98
cachetools==2.0.1
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
cymem==1.31.2
Cython==0.28.2
cytoolz==0.8.2
decorator==4.3.0
dill==0.2.7.1
docutils==0.14
entrypoints==0.2.3
fasttext==0.8.22
gast==0.2.0
gensim==3.4.0
google-api-core==1.1.1
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-translate==1.3.1
googleapis-common-protos==1.5.3
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.0
Keras==2.1.5
kiwisolver==1.0.1
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
numexpr==2.6.4
numpy==1.14.2
pandas==0.22.0
pandocfilters==1.4.2
parso==0.2.0
pathlib==1.0.1
pexpect==4.5.0
pickleshare==0.7.4
plac==0.9.6
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybind11==2.2.2
pycodestyle==2.4.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.4
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
regex==2017.4.5
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.1
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
smart-open==1.5.7
spacy==2.0.11
tables==3.4.2
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==6.10.2
toolz==0.9.0
tornado==5.0.2
tqdm==4.23.0
traitlets==4.3.2
ujson==1.35
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.0.0/en_core_web_sm-2.0.0.tar.gz
https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-2.0.0/es_core_news_sm-2.0.0.tar.gz